+++ Solution/RiskOfChaos/Utilities/Pickup/PickupNotificationFlags.cs
@@ -1,0 +1,11 @@
+﻿namespace RiskOfChaos.Utilities.Pickup
+{
+    public enum PickupNotificationFlags : uint
+    {
+        None = 0,
+        DisplayPushNotification = 1 << 0,
+        DisplayPushNotificationIfNoneQueued = 1 << 1,
+        PlaySound = 1 << 2,
+        SendChatMessage = 1 << 3,
+    }
+}