+++ Solution/RiskOfChaos/Utilities/Pickup/PickupUtils.cs
@@ -6,51 +6,49 @@ namespace RiskOfChaos.Utilities.Pickup
 {
     public static class PickupUtils
     {
-        public static void QueuePickupMessage(CharacterMaster characterMaster, PickupIndex pickupIndex, bool pushNotification, bool playPickupSound)
+        public const PickupNotificationFlags DEFAULT_NOTIFICATION_FLAGS = PickupNotificationFlags.DisplayPushNotification | PickupNotificationFlags.PlaySound | PickupNotificationFlags.SendChatMessage;
+
+        public static void QueuePickupMessage(CharacterMaster characterMaster, PickupIndex pickupIndex, PickupNotificationFlags notificationFlags = DEFAULT_NOTIFICATION_FLAGS)
         {
             PickupsNotificationMessage pickupMessage = new PickupsNotificationMessage(characterMaster, [pickupIndex])
             {
-                DisplayPushNotification = pushNotification,
-                PlaySound = playPickupSound
+                Flags = notificationFlags
             };
 
             NetworkMessageQueue.EnqueueMessage(pickupMessage, NetworkDestination.Clients);
         }
 
-        public static void QueuePickupMessage(string messageToken, PickupIndex pickupIndex, uint pickupQuantity, bool pushNotification, bool playPickupSound)
+        public static void QueuePickupMessage(string messageToken, PickupIndex pickupIndex, uint pickupQuantity, PickupNotificationFlags notificationFlags = DEFAULT_NOTIFICATION_FLAGS)
         {
             PickupsNotificationMessage pickupMessage = new PickupsNotificationMessage(messageToken, [pickupIndex], [pickupQuantity])
             {
-                DisplayPushNotification = pushNotification,
-                PlaySound = playPickupSound
+                Flags = notificationFlags
             };
 
             NetworkMessageQueue.EnqueueMessage(pickupMessage, NetworkDestination.Clients);
         }
 
-        public static void QueuePickupsMessage(CharacterMaster characterMaster, PickupIndex[] pickupIndices, bool pushNotification, bool playPickupSound)
+        public static void QueuePickupsMessage(CharacterMaster characterMaster, PickupIndex[] pickupIndices, PickupNotificationFlags notificationFlags = DEFAULT_NOTIFICATION_FLAGS)
         {
             if (pickupIndices.Length == 0)
                 return;
 
             PickupsNotificationMessage pickupMessage = new PickupsNotificationMessage(characterMaster, pickupIndices)
             {
-                DisplayPushNotification = pushNotification,
-                PlaySound = playPickupSound
+                Flags = notificationFlags
             };
 
             NetworkMessageQueue.EnqueueMessage(pickupMessage, NetworkDestination.Clients);
         }
 
-        public static void QueuePickupsMessage(string messageToken, PickupIndex[] pickupIndices, uint[] pickupQuantities, bool pushNotification, bool playPickupSound)
+        public static void QueuePickupsMessage(string messageToken, PickupIndex[] pickupIndices, uint[] pickupQuantities, PickupNotificationFlags notificationFlags = DEFAULT_NOTIFICATION_FLAGS)
         {
             if (pickupIndices.Length == 0)
                 return;
 
             PickupsNotificationMessage pickupMessage = new PickupsNotificationMessage(messageToken, pickupIndices, pickupQuantities)
             {
-                DisplayPushNotification = pushNotification,
-                PlaySound = playPickupSound
+                Flags = notificationFlags
             };
 
             NetworkMessageQueue.EnqueueMessage(pickupMessage, NetworkDestination.Clients);